+++ src/AutoRest.CSharp/Mgmt/AutoRest/MgmtOutputLibrary.cs
@@ -11,6 +11,7 @@ using AutoRest.CSharp.Common.Utilities;
 using AutoRest.CSharp.Generation.Types;
 using AutoRest.CSharp.Input;
 using AutoRest.CSharp.Mgmt.Decorator;
+using AutoRest.CSharp.Mgmt.Decorator.Transformer;
 using AutoRest.CSharp.Mgmt.Models;
 using AutoRest.CSharp.Mgmt.Output;
 using AutoRest.CSharp.Output.Builders;
@@ -90,8 +91,7 @@ namespace AutoRest.CSharp.Mgmt.AutoRest
 
         public MgmtOutputLibrary()
         {
-            OmitOperationGroups.RemoveOperationGroups();
-            MgmtContext.CodeModel.UpdateSubscriptionIdForAllResource();
+            CodeModelTransformer.Transform();
             _operationGroupToRequestPaths = new Dictionary<OperationGroup, IEnumerable<string>>();
             RawRequestPathToOperationSets = new CachedDictionary<string, OperationSet>(CategorizeOperationGroups);
             OperationsToOperationGroups = new CachedDictionary<Operation, OperationGroup>(PopulateOperationsToOperationGroups);
@@ -110,11 +110,8 @@ namespace AutoRest.CSharp.Mgmt.AutoRest
             _mergedOperations = Configuration.MgmtConfiguration.MergeOperations
                 .SelectMany(kv => kv.Value.Select(v => (FullOperationName: v, MethodName: kv.Key)))
                 .ToDictionary(kv => kv.FullOperationName, kv => kv.MethodName);
-            MgmtContext.CodeModel.VerifyApiVersions();
-            MgmtContext.CodeModel.UpdateAcronyms();
+
             _allSchemas = MgmtContext.CodeModel.AllSchemas;
-            UrlToUri.UpdateSuffix(_allSchemas);
-            MgmtContext.CodeModel.UpdatePatchOperations();
             _allSchemas.VerifyAndUpdateFrameworkTypes();
             _allSchemas.UpdateSealChoiceTypes();
             CommonSingleWordModels.Update(_allSchemas);
