+++ src/AutoRest.CSharp/Mgmt/Decorator/CodeModelExtension.cs
@@ -45,22 +45,6 @@ namespace AutoRest.CSharp.Mgmt.Decorator
             }).ToList();
         }
 
-        public static void UpdatePatchOperations(this CodeModel codeModel)
-        {
-            foreach (var operationGroup in codeModel.OperationGroups)
-            {
-                foreach (var operation in operationGroup.Operations)
-                {
-                    if (operation.GetHttpMethod() == HttpMethod.Patch)
-                    {
-                        var bodyParameter = operation.GetBodyParameter();
-                        if (bodyParameter != null)
-                            bodyParameter.Required = true;
-                    }
-                }
-            }
-        }
-
         public static void VerifyAndUpdateFrameworkTypes(this IEnumerable<Schema> allSchemas)
         {
             foreach (var schema in allSchemas)
@@ -153,164 +137,5 @@ namespace AutoRest.CSharp.Mgmt.Decorator
             return String.Join("", newPropertyWords);
         }
 
-        public static void UpdateSubscriptionIdForAllResource(this CodeModel codeModel)
-        {
-            bool setSubParam = false;
-            foreach (var operationGroup in codeModel.OperationGroups)
-            {
-                foreach (var op in operationGroup.Operations)
-                {
-                    foreach (var p in op.Parameters)
-                    {
-                        // update the first subscriptionId parameter to be 'method' parameter
-                        if (!setSubParam && p.Language.Default.Name.Equals("subscriptionId", StringComparison.OrdinalIgnoreCase))
-                        {
-                            setSubParam = true;
-                            p.Implementation = ImplementationLocation.Method;
-                        }
-                        // update the apiVersion parameter to be 'client' method
-                        if (p.Language.Default.Name.Equals("apiVersion", StringComparison.OrdinalIgnoreCase))
-                        {
-                            p.Implementation = ImplementationLocation.Client;
-                        }
-                    }
-                }
-            }
-        }
-
-        public static void UpdateAcronym(Schema schema)
-        {
-            if (Configuration.MgmtConfiguration.RenameRules.Count == 0)
-                return;
-            var transformer = new NameTransformer(Configuration.MgmtConfiguration.RenameRules);
-            var wordCache = new ConcurrentDictionary<string, string>();
-            UpdateAcronyms(new Schema[] { schema }, transformer, wordCache);
-        }
-
-        public static void UpdateAcronyms(this CodeModel codeModel)
-        {
-            if (Configuration.MgmtConfiguration.RenameRules.Count == 0)
-                return;
-            var transformer = new NameTransformer(Configuration.MgmtConfiguration.RenameRules);
-            var wordCache = new ConcurrentDictionary<string, string>();
-            // first transform all the name of schemas, properties
-            UpdateAcronyms(codeModel.AllSchemas, transformer, wordCache);
-            // transform all the parameter names
-            UpdateAcronyms(codeModel.OperationGroups, transformer, wordCache);
-        }
-
-        private static void UpdateAcronyms(IEnumerable<Schema> allSchemas, NameTransformer transformer, ConcurrentDictionary<string, string> wordCache)
-        {
-            foreach (var schema in allSchemas)
-            {
-                TransformSchema(schema, transformer, wordCache);
-            }
-        }
-
-        private static void UpdateAcronyms(IEnumerable<OperationGroup> operationGroups, NameTransformer transformer, ConcurrentDictionary<string, string> wordCache)
-        {
-            foreach (var operationGroup in operationGroups)
-            {
-                foreach (var operation in operationGroup.Operations)
-                {
-                    TransformOperation(operation, transformer, wordCache);
-                }
-            }
-        }
-
-        private static void TransformOperation(Operation operation, NameTransformer transformer, ConcurrentDictionary<string, string> wordCache)
-        {
-            TransformLanguage(operation.Language, transformer, wordCache);
-            // this iteration only applies to path and query parameter (maybe headers?) but not to body parameter
-            foreach (var parameter in operation.Parameters)
-            {
-                TransformLanguage(parameter.Language, transformer, wordCache);
-            }
-
-            // we need to iterate over the parameters in each request (actually only one request) to ensure the name of body parameters are also taken care of
-            foreach (var request in operation.Requests)
-            {
-                foreach (var parameter in request.Parameters)
-                {
-                    TransformLanguage(parameter.Language, transformer, wordCache);
-                }
-            }
-        }
-
-        private static void TransformSchema(Schema schema, NameTransformer transformer, ConcurrentDictionary<string, string> wordCache)
-        {
-            switch (schema)
-            {
-                case ChoiceSchema choiceSchema:
-                    TransformChoiceSchema(choiceSchema.Language, choiceSchema.Choices, transformer, wordCache);
-                    break;
-                case SealedChoiceSchema sealedChoiceSchema:
-                    TransformChoiceSchema(sealedChoiceSchema.Language, sealedChoiceSchema.Choices, transformer, wordCache);
-                    break;
-                case ObjectSchema objSchema: // GroupSchema inherits ObjectSchema, therefore we do not need to handle that
-                    TransformObjectSchema(objSchema, transformer, wordCache);
-                    break;
-                default:
-                    throw new InvalidOperationException($"Unknown schema type {schema.GetType()}");
-            }
-        }
-
-        private static void TransformChoiceSchema(Languages languages, ICollection<ChoiceValue> choiceValues, NameTransformer transformer, ConcurrentDictionary<string, string> wordCache)
-        {
-            TransformLanguage(languages, transformer, wordCache);
-            TransformChoices(choiceValues, transformer, wordCache);
-        }
-
-        private static void TransformChoices(ICollection<ChoiceValue> choiceValues, NameTransformer transformer, ConcurrentDictionary<string, string> wordCache)
-        {
-            foreach (var choiceValue in choiceValues)
-            {
-                TransformLanguage(choiceValue.Language, transformer, wordCache);
-            }
-        }
-
-        private static void TransformLanguage(Languages languages, NameTransformer transformer, ConcurrentDictionary<string, string> wordCache)
-        {
-            var originalName = languages.Default.Name;
-            if (wordCache.TryGetValue(originalName, out var result))
-            {
-                languages.Default.Name = result;
-                return;
-            }
-            result = transformer.EnsureNameCase(originalName);
-            languages.Default.Name = result;
-            wordCache.TryAdd(originalName, result);
-        }
-
-        private static void TransformObjectSchema(ObjectSchema objSchema, NameTransformer transformer, ConcurrentDictionary<string, string> wordCache)
-        {
-            // transform the name of this schema
-            TransformLanguage(objSchema.Language, transformer, wordCache);
-            foreach (var property in objSchema.Properties)
-            {
-                TransformLanguage(property.Language, transformer, wordCache);
-            }
-        }
-
-        public static void VerifyApiVersions(this CodeModel codeModel)
-        {
-            foreach (var operationGroup in codeModel.OperationGroups)
-            {
-                VerifyApiVersionsWithinOperationGroup(operationGroup);
-            }
-        }
-
-        // Operations within an operation group should use the same API version.
-        // TODO: this might be able to be removed after https://github.com/Azure/autorest.csharp/issues/1917 is resolved.
-        private static void VerifyApiVersionsWithinOperationGroup(OperationGroup operationGroup)
-        {
-            var apiVersionValues = operationGroup.Operations
-                .SelectMany(op => op.Parameters.Where(p => p.Origin == "modelerfour:synthesized/api-version").Select(p => ((ConstantSchema)p.Schema).Value.Value))
-                .ToHashSet();
-            if (apiVersionValues.Count > 1)
-            {
-                throw new InvalidOperationException($"Multiple api-version values found in the operation group: {operationGroup.Key}. Please rename the operation group for some operations so that all operations in one operation group share the same API version.");
-            }
-        }
     }
 }
