+++ src/AutoRest.CSharp/Mgmt/Decorator/Transformer/BodyParameter.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using AutoRest.CSharp.Input;
+using AutoRest.CSharp.Mgmt.AutoRest;
+
+namespace AutoRest.CSharp.Mgmt.Decorator.Transformer
+{
+    internal static class BodyParameter
+    {
+    }
+}
