+++ src/AutoRest.CSharp/Mgmt/Decorator/Transformer/OmitOperationGroups.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using AutoRest.CSharp.Input;
 using AutoRest.CSharp.Mgmt.AutoRest;
 
-namespace AutoRest.CSharp.Mgmt.Decorator
+namespace AutoRest.CSharp.Mgmt.Decorator.Transformer
 {
     internal static class OmitOperationGroups
     {
